fixing countryDetail
